refactor: security feature moved in avenirs-portfolio-common
