--- src/main/java/fr/avenirsesr/portfolio/security/infrastructure/model/HmacAuthenticationToken.java
@@ -1,26 +1,0 @@
-package fr.avenirsesr.portfolio.security.infrastructure.model;
-
-import java.util.Collections;
-import java.util.UUID;
-import org.springframework.security.authentication.AbstractAuthenticationToken;
-
-public class HmacAuthenticationToken extends AbstractAuthenticationToken {
-
-  private final UUID userId;
-
-  public HmacAuthenticationToken(UUID userId) {
-    super(Collections.emptyList());
-    this.userId = userId;
-    setAuthenticated(true);
-  }
-
-  @Override
-  public Object getCredentials() {
-    return "";
-  }
-
-  @Override
-  public Object getPrincipal() {
-    return userId;
-  }
-}
