--- src/test/java/fr/avenirsesr/portfolio/security/infrastructure/filter/HmacAuthenticationFilterTest.java
@@ -1,380 +1,0 @@
-package fr.avenirsesr.portfolio.security.infrastructure.filter;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.junit.jupiter.api.Assertions.fail;
-import static org.mockito.Mockito.mockStatic;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-import fr.avenirsesr.portfolio.common.testutils.BddLogger;
-import fr.avenirsesr.portfolio.security.infrastructure.model.enums.ESecurityKeys;
-import fr.avenirsesr.portfolio.user.domain.exception.UserNotAuthorizedException;
-import fr.avenirsesr.portfolio.user.infrastructure.adapter.model.UserPayload;
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import java.nio.charset.StandardCharsets;
-import java.time.Instant;
-import java.util.Base64;
-import java.util.UUID;
-import javax.crypto.Mac;
-import javax.crypto.spec.SecretKeySpec;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.Spy;
-import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.test.util.ReflectionTestUtils;
-
-@ExtendWith(MockitoExtension.class)
-class HmacAuthenticationFilterTest {
-
-  @Spy @InjectMocks private HmacAuthenticationFilter filter;
-
-  @Mock private HttpServletRequest request;
-
-  @Mock private HttpServletResponse response;
-
-  @Mock private FilterChain filterChain;
-
-  private ObjectMapper objectMapper;
-  private static final String TEST_KEY = "TEST_KEY";
-  private static final String TEST_SECRET = "test-secret";
-  private static final UUID TEST_UUID = UUID.fromString("123e4567-e89b-12d3-a456-426614174000");
-
-  @BeforeEach
-  void setUp() {
-    SecurityContextHolder.clearContext();
-    objectMapper = new ObjectMapper();
-    objectMapper.registerModule(new JavaTimeModule());
-    ReflectionTestUtils.setField(
-        filter,
-        "permitAllPathsString",
-        "/avenirs-portfolio-api/swagger-ui/**,/avenirs-portfolio-api/api-docs/**,/favicon.ico,/actuator/health");
-  }
-
-  @Test
-  void shouldNotFilterPublicPaths() throws ServletException {
-    BddLogger.given("a request on any public path");
-    BddLogger.when("performing the request");
-    BddLogger.then("it should not filter");
-    when(request.getRequestURI()).thenReturn("/avenirs-portfolio-api/swagger-ui/index.html");
-    assertTrue(filter.shouldNotFilter(request));
-
-    when(request.getRequestURI()).thenReturn("/avenirs-portfolio-api/api-docs");
-    assertTrue(filter.shouldNotFilter(request));
-
-    when(request.getRequestURI()).thenReturn("/favicon.ico");
-    assertTrue(filter.shouldNotFilter(request));
-
-    when(request.getRequestURI()).thenReturn("/actuator/health");
-    assertTrue(filter.shouldNotFilter(request));
-
-    when(request.getRequestURI()).thenReturn("/api/protected");
-    assertFalse(filter.shouldNotFilter(request));
-  }
-
-  @Test
-  void shouldNotFilterPublicPath() {
-    BddLogger.given("a request on any public path");
-    BddLogger.when("performing the request");
-    BddLogger.then("it should not filter");
-    try {
-      ReflectionTestUtils.setField(filter, "permitAllPathsString", "/public/path,/another/path");
-      ReflectionTestUtils.setField(
-          filter, "permitAllPathsList", java.util.Arrays.asList("/public/path", "/another/path"));
-
-      when(request.getRequestURI()).thenReturn("/public/path/resource");
-
-      assertTrue(filter.shouldNotFilter(request));
-    } catch (Exception e) {
-      fail("Test setup failed: " + e.getMessage());
-    }
-  }
-
-  @Test
-  void shouldFilterNonPublicPath() {
-    BddLogger.given("a request on any non public path");
-    BddLogger.when("performing the request");
-    BddLogger.then("it should filter");
-    try {
-      ReflectionTestUtils.setField(filter, "permitAllPathsString", "/public/path,/another/path");
-      ReflectionTestUtils.setField(
-          filter, "permitAllPathsList", java.util.Arrays.asList("/public/path", "/another/path"));
-
-      when(request.getRequestURI()).thenReturn("/protected/resource");
-
-      assertFalse(filter.shouldNotFilter(request));
-    } catch (Exception e) {
-      fail("Test setup failed: " + e.getMessage());
-    }
-  }
-
-  @Test
-  void shouldAuthenticateWithValidSignature() throws ServletException, IOException {
-    BddLogger.given("a request");
-    UserPayload userPayload = new UserPayload();
-    userPayload.setSub(TEST_UUID);
-    userPayload.setExp(Instant.now().plusSeconds(3600));
-
-    String payload = objectMapper.writeValueAsString(userPayload);
-    String signature = generateHmacSignature(payload);
-
-    when(request.getHeader("X-Context-Kid")).thenReturn(TEST_KEY);
-    when(request.getHeader("X-Context-Signature")).thenReturn(signature);
-    when(request.getHeader("X-Signed-Context")).thenReturn(payload);
-
-    BddLogger.when("performing the request with valid signature");
-    BddLogger.then("it should authenticate");
-    try (var mockedStatic = mockStatic(ESecurityKeys.class)) {
-      mockedStatic.when(() -> ESecurityKeys.getSecretByKey(TEST_KEY)).thenReturn(TEST_SECRET);
-
-      filter.doFilterInternal(request, response, filterChain);
-
-      verify(filterChain).doFilter(request, response);
-      assertNotNull(SecurityContextHolder.getContext().getAuthentication());
-      assertEquals(
-          TEST_UUID, SecurityContextHolder.getContext().getAuthentication().getPrincipal());
-    }
-  }
-
-  @Test
-  void shouldRejectExpiredPayload() {
-    BddLogger.given("a request");
-    UserPayload userPayload = new UserPayload();
-    userPayload.setSub(TEST_UUID);
-    userPayload.setExp(Instant.now().minusSeconds(3600));
-
-    BddLogger.when("performing the request with expired payload");
-    BddLogger.then("it should reject it");
-    try {
-      String payload = objectMapper.writeValueAsString(userPayload);
-      String signature = generateHmacSignature(payload);
-
-      when(request.getHeader("X-Context-Kid")).thenReturn(TEST_KEY);
-      when(request.getHeader("X-Context-Signature")).thenReturn(signature);
-      when(request.getHeader("X-Signed-Context")).thenReturn(payload);
-
-      try (var mockedStatic = mockStatic(ESecurityKeys.class)) {
-        mockedStatic.when(() -> ESecurityKeys.getSecretByKey(TEST_KEY)).thenReturn(TEST_SECRET);
-
-        assertThrows(
-            UserNotAuthorizedException.class,
-            () -> filter.doFilterInternal(request, response, filterChain));
-      }
-    } catch (Exception e) {
-      fail("Test setup failed: " + e.getMessage());
-    }
-  }
-
-  @Test
-  void shouldRejectInvalidSignature() {
-    BddLogger.given("a request");
-    UserPayload userPayload = new UserPayload();
-    userPayload.setSub(TEST_UUID);
-    userPayload.setExp(Instant.now().plusSeconds(3600));
-
-    BddLogger.when("performing the request with invalid signature");
-    BddLogger.then("it should reject it");
-    try {
-      String payload = objectMapper.writeValueAsString(userPayload);
-      String invalidSignature = "invalid-signature";
-
-      when(request.getHeader("X-Context-Kid")).thenReturn(TEST_KEY);
-      when(request.getHeader("X-Context-Signature")).thenReturn(invalidSignature);
-      when(request.getHeader("X-Signed-Context")).thenReturn(payload);
-
-      try (var mockedStatic = mockStatic(ESecurityKeys.class)) {
-        mockedStatic.when(() -> ESecurityKeys.getSecretByKey(TEST_KEY)).thenReturn(TEST_SECRET);
-
-        assertThrows(
-            UserNotAuthorizedException.class,
-            () -> filter.doFilterInternal(request, response, filterChain));
-      }
-    } catch (Exception e) {
-      fail("Test setup failed: " + e.getMessage());
-    }
-  }
-
-  @Test
-  void shouldHandleExceptionDuringSignatureVerification() {
-    BddLogger.given("a request");
-    UserPayload userPayload = new UserPayload();
-    userPayload.setSub(TEST_UUID);
-    userPayload.setExp(Instant.now().plusSeconds(3600));
-
-    BddLogger.when("performing the request and an exception occurs during signature verification");
-    BddLogger.then("it should handle it");
-    try {
-      String payload = objectMapper.writeValueAsString(userPayload);
-
-      when(request.getHeader("X-Context-Kid")).thenReturn(TEST_KEY);
-      when(request.getHeader("X-Context-Signature")).thenReturn("valid-looking-signature");
-      when(request.getHeader("X-Signed-Context")).thenReturn(payload);
-
-      try (var mockedStatic = mockStatic(ESecurityKeys.class)) {
-        mockedStatic.when(() -> ESecurityKeys.getSecretByKey(TEST_KEY)).thenReturn(null);
-
-        assertThrows(
-            UserNotAuthorizedException.class,
-            () -> filter.doFilterInternal(request, response, filterChain));
-      }
-    } catch (Exception e) {
-      fail("Test setup failed: " + e.getMessage());
-    }
-  }
-
-  @Test
-  void shouldRejectNullPayload() {
-    BddLogger.given("a request");
-    BddLogger.when("performing the request with null payload");
-    BddLogger.then("it should reject it");
-    try {
-      when(request.getHeader("X-Context-Kid")).thenReturn(TEST_KEY);
-      when(request.getHeader("X-Context-Signature")).thenReturn("some-signature");
-      when(request.getHeader("X-Signed-Context")).thenReturn("invalid-json");
-
-      try (var mockedStatic = mockStatic(ESecurityKeys.class)) {
-        mockedStatic.when(() -> ESecurityKeys.getSecretByKey(TEST_KEY)).thenReturn(TEST_SECRET);
-
-        assertThrows(
-            Exception.class, () -> filter.doFilterInternal(request, response, filterChain));
-      }
-    } catch (Exception e) {
-      fail("Test setup failed: " + e.getMessage());
-    }
-  }
-
-  @Test
-  void shouldRejectNullPayloadString() {
-    BddLogger.given("a request");
-    BddLogger.when("performing the request with null payload string");
-    BddLogger.then("it should reject it");
-    try {
-      when(request.getHeader("X-Context-Kid")).thenReturn(TEST_KEY);
-      when(request.getHeader("X-Context-Signature")).thenReturn("some-signature");
-      when(request.getHeader("X-Signed-Context")).thenReturn(null);
-
-      try (var mockedStatic = mockStatic(ESecurityKeys.class)) {
-        mockedStatic.when(() -> ESecurityKeys.getSecretByKey(TEST_KEY)).thenReturn(TEST_SECRET);
-
-        assertThrows(
-            Exception.class, () -> filter.doFilterInternal(request, response, filterChain));
-      }
-    } catch (Exception e) {
-      fail("Test setup failed: " + e.getMessage());
-    }
-  }
-
-  @Test
-  void shouldRejectMissingSecretKey() {
-    BddLogger.given("a request");
-    UserPayload userPayload = new UserPayload();
-    userPayload.setSub(TEST_UUID);
-    userPayload.setExp(Instant.now().plusSeconds(3600));
-
-    BddLogger.when("performing the request without secret key");
-    BddLogger.then("it should reject it");
-    try {
-      String payload = objectMapper.writeValueAsString(userPayload);
-      String signature = generateHmacSignature(payload);
-
-      when(request.getHeader("X-Context-Kid")).thenReturn("UNKNOWN_KEY");
-      when(request.getHeader("X-Context-Signature")).thenReturn(signature);
-      when(request.getHeader("X-Signed-Context")).thenReturn(payload);
-
-      try (var mockedStatic = mockStatic(ESecurityKeys.class)) {
-        mockedStatic.when(() -> ESecurityKeys.getSecretByKey("UNKNOWN_KEY")).thenReturn(null);
-
-        assertThrows(
-            UserNotAuthorizedException.class,
-            () -> filter.doFilterInternal(request, response, filterChain));
-      }
-    } catch (Exception e) {
-      fail("Test setup failed: " + e.getMessage());
-    }
-  }
-
-  @Test
-  void shouldRejectMissingSignature() {
-    BddLogger.given("a request");
-    UserPayload userPayload = new UserPayload();
-    userPayload.setSub(TEST_UUID);
-    userPayload.setExp(Instant.now().plusSeconds(3600));
-
-    BddLogger.when("performing the request without signature");
-    BddLogger.then("it should reject it");
-    try {
-      String payload = objectMapper.writeValueAsString(userPayload);
-
-      when(request.getHeader("X-Context-Kid")).thenReturn(TEST_KEY);
-      when(request.getHeader("X-Context-Signature")).thenReturn(null);
-      when(request.getHeader("X-Signed-Context")).thenReturn(payload);
-
-      try (var mockedStatic = mockStatic(ESecurityKeys.class)) {
-        mockedStatic.when(() -> ESecurityKeys.getSecretByKey(TEST_KEY)).thenReturn(TEST_SECRET);
-
-        assertThrows(
-            UserNotAuthorizedException.class,
-            () -> filter.doFilterInternal(request, response, filterChain));
-      }
-    } catch (Exception e) {
-      fail("Test setup failed: " + e.getMessage());
-    }
-  }
-
-  @Test
-  void shouldRejectPayloadWithNullExpiration() {
-    BddLogger.given("a request");
-    UserPayload userPayload = new UserPayload();
-    userPayload.setSub(TEST_UUID);
-    userPayload.setExp(null);
-
-    BddLogger.when("performing the request without expiration");
-    BddLogger.then("it should reject it");
-    try {
-      String payload = objectMapper.writeValueAsString(userPayload);
-      String signature = generateHmacSignature(payload);
-
-      when(request.getHeader("X-Context-Kid")).thenReturn(TEST_KEY);
-      when(request.getHeader("X-Context-Signature")).thenReturn(signature);
-      when(request.getHeader("X-Signed-Context")).thenReturn(payload);
-
-      try (var mockedStatic = mockStatic(ESecurityKeys.class)) {
-        mockedStatic.when(() -> ESecurityKeys.getSecretByKey(TEST_KEY)).thenReturn(TEST_SECRET);
-
-        assertThrows(
-            Exception.class, () -> filter.doFilterInternal(request, response, filterChain));
-      }
-    } catch (Exception e) {
-      fail("Test setup failed: " + e.getMessage());
-    }
-  }
-
-  private String generateHmacSignature(String payload) {
-    try {
-      Mac sha256Hmac = Mac.getInstance("HmacSHA256");
-      SecretKeySpec secretKeySpec =
-          new SecretKeySpec(
-              HmacAuthenticationFilterTest.TEST_SECRET.getBytes(StandardCharsets.UTF_8),
-              "HmacSHA256");
-      sha256Hmac.init(secretKeySpec);
-
-      byte[] signedBytes = sha256Hmac.doFinal(payload.getBytes(StandardCharsets.UTF_8));
-      return Base64.getEncoder().encodeToString(signedBytes);
-    } catch (Exception e) {
-      throw new RuntimeException("Failed to generate HMAC signature", e);
-    }
-  }
-}
